This is the first commit of everything that Omer and I have done.   Not commented yet.
I also made it so Subversion would ignore class files in this directory, via the propset command.  This means that when you check the status of your working copy, via the 'svn status' command it won't be constantly telling you about the *.class files that you have changed or not added to the repository.

-Benji