Did a lot of work on StartUp. It's a GUI to get into the game. It can take all the command line arguments of both the Server and Representation3D. But versus having to know each of those arguments you just need to select a check box for those arguments.
Apparently with the latest revision though the minimum requirements needed to run the game have increased a little bit. I wasn't able to run the game with my current laptop graphics processor, a little integrated Intel chip that can't really run too much anyway. On the other hand my Desktop running an ATI x800 is able to run it just fine. My thoughts are OpenGL 1.2 or higher looking briefly at the error output on the command line. 

Back to StartUp.java I was wondering if anyone had any ideas on making it so the output of the Server and/or Representation3D can go to a window on the StartUp or even its own tab. 

Next step would be to get it so it can be run from an executable JAR file or even an applet runnable in a browser. But if the system requirements have changed it might be better to not run in a browser. 

Enough ranting
--Guillermo