Modified StartUp a little bit more. Now it runs things like a charm with a splash screen for both the Server and the Client. Added a window to the server so it can be closed without going into a task manager or using a kill command. The images folder is where we can store the images to use for the splash screens.
Also please check the following locations for warnings: 
ClientThread, line 9, verbose is never read locally
Resolver, Line 6, ruleFactory is never read locally
AttributesHashMap, line 31, entry is never read
ElementFactory, line 48, plural is never read
ElementFactory, line 49, facingNodes is never read
GameElementBranch, line 316, rotationAlpha2 is never read
THE BIG TWO TO CHECK AND FIX THE WARNING
Resolver, line 148, TYPE SAFETY: The expression of type IncrementedArray[] needs unchecked conversion to conform to incrementedArray<GameElement>[]
ElementStack, line 66, Type safety: The expression of type LinkedHolder needs unchecked conversion to conform to the LinkedHolder<K>

The other files that were modified had other warnings that I was able to take a look at and see they were not used at all. But you can check to make sure I didn't take something out that was necessary. 

I added the console class for future work so we can have a place to output all the messages from the server and/or the client. Right now it's just a shell to be filled in. If I was reading things correctly Java 6, there is a console with a few methods ready to be implemented. The direct link to the API is as follows: http://java.sun.com/javase/6/docs/api/java/io/Console.html 

--Guillermo