Unless there are major changes to either Client, Rep3D, or Server StartUp should be in it's final form, sans comments.
Now for the information on what changed. Added the options for the newly implemented options for text (server) and dir, eext, efiles (client). Changed the variables so its a little easier to read. And when you hit either the Client or Server button you will get a printout of what options were selected. 

For the server if you are using text mode the splash screen won't show up. But if you are going to use the GUI it will pop-up. The client will always use the splash screen. (unless we decide to turn it off/change which image to use)

Maybe we can decide to which image to pick for each application, server or client. I just decided on the Java mascot, Duke, because it seemed appropriate :-D. 

Until Monday
--Guillermo

PS Omer/Benji I can't connect to your server it's saying that the server refused my connection.