We made win and loss conditions (that you find four pots of gold first), and messages get sent to all players whenever anyone finds a pot of gold. When one player finds four pots, everyone is forced to quit.
We also added one new option to the command line start-up for the Client: -n <name>
This allows players to specify a name other than "Player <id>" (which is the default). Ideally, this new field would be added to the StartUp GUI, but given that we're demoing tomorrow, that might not be possible, which is fine.

Also, to make the maze game, we changed some of the rules (like movement), so we stored the old rules in test.ccrold - the new rules-file is maze.ccr

Basically: we should have a working game :)

- Benjamin and Omer.